Rework test suite a bit more, add some moar tests
